Revise architecture
